game-engine/requirements.txt
paho-mqtt==2.1.*